+++ src/main/java/dev/nicho/rolesync/bot/discord/ReplyType.java
@@ -1,0 +1,7 @@
+package dev.nicho.rolesync.bot.discord;
+
+public enum ReplyType {
+    INFO,
+    SUCCESS,
+    ERROR,
+}